+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Add initilize method with dsn to init SentryClient.
+* Make FlutterSentry a Singleton.
+
 ## 0.0.1+2
 
 * Update examples in README.